Update patch set 4
Patch Set 4:

But this isn't 2012 anymore -- if this behavior had been abandoned sooner (when bringing the devices up for CM), you would not be opposing this.

I'll use other OEMs as an example here: stock (OEM) recoveries are useless for people who are flashing CM, so those maintainers have to assume their users are using a (custom) recovery from the recovery partition; why may we not do the same?

It is much safer to assume that the recovery ramdisk on FOTA will always be working than to assume as much about the recovery ramdisk packed into the boot image: Once installed, a working FOTA recovery would not change, but a ramdisk recovery changes with each build, increasing the chances of failure. Of course, you could argue that in those cases, a FOTA recovery would be a backup, but what about the people that do not use a FOTA recovery?

What's infuriating about the ramdisk recovery is that the majority of users do not flash a FOTA recovery, and due to the LZMA patch not being merged (rendering several types of recoveries useless without custom patches that can only be run through recoveries that require the LZMA patch (creating a loop of redundancy)), this causes issues to developers who create flashable zips with scripts that assume certain features are working.

Patch-set: 4