Update patch set 4
Patch Set 4:

I won't disagree with that.

And the other OEMS part too. IIRC we did ask people to flash CM recovery to FOTAKernel, and also provide separate recovery images even if they aren't fully useful. And other users can flash something else and it won't get updated.

The increase in chances of failure argument is good too, but since we've had this since forever we can't assume *all* users have FOTAKernel flashed.

What we can do is use that approach for newer devices from the very first build, that is something I wouldn't disagree with, but we'll need to think a bit more and talk with other maintainers before making that decision.

For the LZMA patch, could you bring it up to 12.1 (with needed changes elsewhere, if any), and let's try to give it another look.

Patch-set: 4