Update patch set 4
Patch Set 4:

We could simply issue out a warning that the next nightly is a "dragons ahead; beware" type of thing and that a recovery is a must.
Or at the very least, get ready to merge the change when CM13 hits -- start issuing warnings about the recovery now, so people have time to prepare, and start anew with Android 6.0.

I'll push the LZMA patch again into review. It's been tried and tested on many devices, and multiple users can confirm it working.

Patch-set: 4