Update patch set 4
Patch Set 4:

I don't really want to do that.

Not everyone would see it, and the current situation works well enough.

Let's get that LZMA patch in and then the user can use any recovery he wants, and if not, there'll always be what we bundle.


For new devices, I'm not opposed to this like I said, but it would still need some thought.

Patch-set: 4