Update patch set 4
Patch Set 4:

Just food for thought, but if all Sony device maintainers agree to apply this to newer devices when those come in, you should then also apply it to devices that came before those new ones (but are still supported, of course), for the sake of sanity and cleanness -- that way, not every Sony device has to be a special snowflake and the phones under the Sony brand can (and actually should) have similar, if not identical setups.

This behavior should however be abandoned regardless when the rest of the 8974 devices get the new bootloader.

Publishing alerts and warnings would not be that difficult (and it should not be -- you are a maintainer of one of the biggest custom ROMs; if you and your team cannot alert users to a certain new feature or danger, no team can -- people react to jerpelea as a god on XDA), and at one point in time this will have to be done, or the whole "new bootloader" project from Sony may well not be there.

Lastly, I'm throwing in a thought from a fellow developer on this patch: "Teach users good habits instead of letting bad behavior accumulate -- in this case, bad behavior meaning not installing a separate recovery."

Patch-set: 4