Update patch set 4
Patch Set 4:

Perhaps that is something we (FXP) can help solve, if we publish an article and provide a set of good recovery images. That way everyone will know something is changing, and will also have easy access to a good recovery.

Perhaps Marshmallow / cm-13.0 will be a good place to make the change, instead of in the middle of cm-12.1. We'd need to get it working now so we can merged it straight into the cm-13.0 branches.

B2G is already using the extract from FOTAKernel exclusively.

Patch-set: 4