Update patch set 4
Patch Set 4:

> Perhaps Marshmallow / cm-13.0 will be a good place to make the
 > change, instead of in the middle of cm-12.1. We'd need to get it
 > working now so we can merged it straight into the cm-13.0 branches.

This is what I recommended earlier as the very least that should be done.
The fact that the FXP team is prepared to cooperate only renders the point of people not seeing the warning moot -- those that do not read announcements from FXP have clearly not done their research, and that is not something we should account for.

Patch-set: 4