Update patch set 4
Patch Set 4:

> you really should account for all possible sitautions and outcomes,
 > not just the ones that make your point

I have thought about this, and the conclusions are all of the comments I've made under this patch. If there is a situation and/or outcome that might not favor the patch (or vice-versa), then by all means please point it out -- gerrit is here to review the mistakes in each other's code, not so much to find errors in each other.

I agree that there might be possible issues with what I'm trying to do, but until I find them or they are pointed out to me, I cannot fix them or defend my stance and my thinking behind them.

Patch-set: 4