Update patch set 4
Patch Set 4:

> Just jotting this down as another thing that should be taken into
 > account: the z3c recently got approved as an official TWRP device.
 > Without this patch, many users report that they cannot enter into
 > TWRP, but are instead forced into using Cyanogen Recovery (despite
 > correctly issuing the fastboot command). I've verified this as true
 > on the z3c, though it may not be true for other 8974 devices,
 > although, at least for shinano, that seems highly unlikely.

One issue picked away (and it took a lot of help from the TWRP crew to get this one); part of the problem is that extract_elf_ramdisk's buffer is too small. Change submitted at http://review.cyanogenmod.org/#/c/106586/ .

Another issue still remains however, where, if extract_elf_ramdisk is not updated with the linked patch, the FOTA recovery will refuse to boot if it's bigger than the buffer, _unless_ the recovery inside the boot image is not present.

Patch-set: 4